+++ tests/test_extended.py
@@ -1,34 +1,16 @@
 from datetime import datetime, timedelta
 
-import pandas as pd
 import pytest
 from dateutil import rrule as rr
 
-from assume.common.base import SupportsMinMax
 from assume.common.market_objects import MarketConfig, MarketProduct
 from assume.strategies import OTCStrategy
-
-
-class MockUnit(SupportsMinMax):
-    def __init__(self, index):
-        super().__init__("", "", "", {}, index, None)
-        self.max_power = 1000
-        self.min_power = 0
-
-    def calculate_min_max_power(
-        self, start: pd.Timestamp, end: pd.Timestamp, product_type="energy"
-    ) -> tuple[pd.Series, pd.Series]:
-        return 0, 400
-
-    def calculate_marginal_cost(self, start: pd.Timestamp, power: float) -> float:
-        return 3
-
 
 start = datetime(2023, 7, 1)
 end = datetime(2023, 7, 2)
 
 
-def test_otc_strategy():
+def test_otc_strategy(mock_supports_minmax):
     strategy = OTCStrategy(scale_firm_power_capacity=0.1)
 
     mc = MarketConfig(
@@ -39,12 +21,7 @@ def test_otc_strategy():
         [MarketProduct(timedelta(hours=1), 1, timedelta(hours=1))],
     )
 
-    index = pd.date_range(
-        start=datetime(2023, 7, 1),
-        end=datetime(2023, 7, 2),
-        freq="1h",
-    )
-    unit = MockUnit(index)
+    unit = mock_supports_minmax
     start = datetime(2023, 7, 1)
     end = datetime(2023, 7, 2)
     product_tuples = [(start, end, None)]
@@ -58,7 +35,7 @@ def test_otc_strategy():
 
 
 @pytest.mark.parametrize("scale", [0.1, 0.2, 1.0, 0.0])
-def test_otc_strategy_scaled(scale):
+def test_otc_strategy_scaled(scale, mock_supports_minmax):
     strategy = OTCStrategy(scale_firm_power_capacity=scale)
 
     mc = MarketConfig(
@@ -69,12 +46,7 @@ def test_otc_strategy_scaled(scale):
         [MarketProduct(timedelta(hours=1), 1, timedelta(hours=1))],
     )
 
-    index = pd.date_range(
-        start=datetime(2023, 7, 1),
-        end=datetime(2023, 7, 2),
-        freq="1h",
-    )
-    unit = MockUnit(index)
+    unit = mock_supports_minmax
     start = datetime(2023, 7, 1)
     end = datetime(2023, 7, 2)
     product_tuples = [(start, end, None)]
